Make navboard wrapper the main executable
navboard-local has been renamed to navboard.sh and this will be
installed to /usr/bin and be executed by the call to navboard. This
means that users custom boards will be available by default without
needing a special command. If the config directory doesn't exist than
the default navboard binary will be invoked instead.