Set READING env var when spawning a command
When we spawn command we either care about its return value/output or we
don't. In the former case we'll often use it to set internal state.

The Key struct allows function pointers to be set to read initial
state and then to run a command on action. However, it only allows a
single string to be specified. Now users can check to see if the new
READING var is present and decided whether to issue the read or write
command.

For example:
```
Key key = { .loadValue=readValueFromCmd, .onDrag = spawnCmd, .arg.s="if [ $READING -eq 0 ]; then echo $KEY_VALUE%; else echo $$; fi"},
```
Which will initialize a slider to the current pid and then output the
value on change.